price + menu fix when scrolling

+++ Utilities/Commons/CommonFunctions.cs
@@ -7,7 +7,7 @@ namespace Utilities.Commons
 {
     public class CommonFunctions
     {
-        public static string FormatNumber(double number, int precision)
+        public static string FormatNumber(decimal number, int precision)
         {
             var a = number.ToString($"N{precision}").Split('.');
             a[0] = a[0].Replace("/\\d(?=(\\d{3})+$)/g", "$&,");
